Add command handling system
