+++ app/src/main/java/io/github/boogiemonster1o1/eyeyoureadyforit/command/Command.java
@@ -1,0 +1,10 @@
+package io.github.boogiemonster1o1.eyeyoureadyforit.command;
+
+import discord4j.core.event.domain.interaction.SlashCommandEvent;
+import io.github.boogiemonster1o1.eyeyoureadyforit.data.ChannelSpecificData;
+import org.reactivestreams.Publisher;
+
+@FunctionalInterface
+public interface Command {
+	Publisher<?> handle(SlashCommandEvent event, ChannelSpecificData csd);
+}
