+++ app/src/main/java/io/github/boogiemonster1o1/eyeyoureadyforit/db/stats/StatisticsManager.java
@@ -13,7 +13,7 @@ import java.util.ArrayList;
 
 public final class StatisticsManager {
 
-	public static Mono initDb(Snowflake guildId) {
+	public static Mono<?> initDb(Snowflake guildId) {
 		return Mono.fromRunnable(() -> {
 			DataSource.get().withExtension(DataDao.class, dao -> {
 				dao.createTable(guildId.asString());
